Add test for encoding RC options
