+++ modules/polling/context/BallotContext.tsx
@@ -13,8 +13,7 @@ import React, { ReactNode, useEffect, useState } from 'react';
 import { toast } from 'react-toastify';
 import shallow from 'zustand/shallow';
 import { Ballot, BallotVote } from '../types/ballot';
-import { fromBuffer } from 'lib/utils';
-
+import { parsePollOptions } from '../helpers/parsePollOptions';
 interface ContextProps {
   ballot: Ballot;
   transaction?: Transaction;
@@ -179,19 +178,12 @@ export const BallotProvider = ({ children }: PropTypes): React.ReactElement => {
       }
     });
 
-    const optionIds = pollOptions.map(option => {
-      if (!Array.isArray(option)) return option;
-      if (option.length === 1) return option[0];
-      const byteArray = new Uint8Array(32);
-      option.forEach((optionIndex, i) => {
-        byteArray[byteArray.length - i - 1] = optionIndex;
-      });
-      return fromBuffer(byteArray, {}).toString();
-    });
+    const optionIds = parsePollOptions(pollOptions);
 
     const voteTxCreator = voteDelegateContract
       ? () => voteDelegateContract['votePoll(uint256[],uint256[])'](pollIds, optionIds)
-      : () => polling['vote(uint256[],uint256[])'](pollIds, optionIds);
+      : // vote with array arguments can be used for single vote and multiple vote
+        () => polling['vote(uint256[],uint256[])'](pollIds, optionIds);
 
     const txId = track(voteTxCreator, account, `Voting on ${Object.keys(ballot).length} polls`, {
       pending: txHash => {
